+++ src/tokens.h
@@ -35,13 +35,10 @@
 #define STORE           '!'
 #define LOAD            ';'
 
-#define OUT             '.'
-#define IN              ','
 #define MEM_JUMP        'm' //signed int // unused
 #define COPY_FROM       'v' // unused
 #define RUN             '^' // unused
 #define CODE_JMP        'j' // unused
-
 
 #define NEW_VAR         '^'